app/src/main/java/com/puzzlebench/katarobolectric/MainActivity.java
package com.puzzlebench.katarobolectric;

import android.support.v7.app.AppCompatActivity;
import android.os.Bundle;
import android.widget.TextView;

public class MainActivity extends AppCompatActivity {

    TextView textView;
    @Override
    protected void onCreate(Bundle savedInstanceState) {
        super.onCreate(savedInstanceState);
        setContentView(R.layout.activity_main);
        textView = (TextView)findViewById(R.id.texview);
        textView.setText("H");
    }

    public void getText()
    {
        textView.getText().toString();

    }
}
